matching cost done

+++ include/types.hpp
@@ -7,8 +7,8 @@
 
 namespace sgm {
 
-using feature_type = uint32_t;
-using cost_type = uint8_t;
+using feature_type = uint64_t;
+using cost_type = int8_t;
 using cost_sum_type = uint16_t;
 using output_type = uint16_t;
